fix: JaxProtection

+++ contracts/JaxProtection.sol
@@ -22,13 +22,13 @@ contract JaxProtection {
         if(data_hash != protection.data_hash || protection.sender != msg.sender) {
         protection.sender = msg.sender;
         protection.data_hash = keccak256(msg.data);
-        protection.request_timestamp = uint8(block.timestamp);
-        protection.executed = false;
-        emit Request_Update(msg.sig, msg.data);
-        return;
+            protection.request_timestamp = uint8(block.timestamp);
+            protection.executed = false;
+            emit Request_Update(msg.sig, msg.data);
+            return;
         }
         require(protection.executed == false, "Already executed");
-        require(block.timestamp >= protection.request_timestamp + 1 minutes, "Running is Locked");
+        require(block.timestamp >= uint(protection.request_timestamp) + 1 minutes, "Running is Locked");
         _;
         protection.executed = true;
     }
